Backend/requirements.txt
# CareerPilot Backend — Python dependencies
# Generated for Django 5.x + Python 3.11

# Core framework
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0

# CORS
django-cors-headers>=4.3,<5.0

# Environment variables
python-dotenv>=1.0,<2.0

# Database URL parsing (for Postgres in production)
dj-database-url>=2.1,<4.0

# Production WSGI server
gunicorn>=21.0,<23.0

# Static files
whitenoise>=6.6,<7.0

# PostgreSQL Database Driver (for Render)
psycopg2-binary>=2.9,<3.0

# AI integrations
google-genai>=1.0,<2.0
